Split connection clause into different files
