+++ packages/graphql/src/translate/connection-clause/get-sort-fields.ts
@@ -1,0 +1,42 @@
+/*
+ * Copyright (c) "Neo4j"
+ * Neo4j Sweden AB [http://neo4j.com]
+ *
+ * This file is part of Neo4j.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import type { ResolveTree } from "graphql-parse-resolve-info";
+import type { ConnectionSortArg, GraphQLSortArg } from "../../types";
+
+type SortFields = {
+    edge: GraphQLSortArg;
+    node: GraphQLSortArg;
+};
+
+/** Returns sort fields for connections, split between edge and node */
+export function getSortFields(resolveTree: ResolveTree): SortFields {
+    const sortInput = (resolveTree.args.sort ?? []) as ConnectionSortArg[];
+
+    return {
+        edge: getSortFieldsByElement(sortInput, "edge"),
+        node: getSortFieldsByElement(sortInput, "node"),
+    };
+}
+
+function getSortFieldsByElement(sortInput: ConnectionSortArg[], element: "node" | "edge"): GraphQLSortArg {
+    return sortInput.reduce((acc, f) => {
+        return { ...acc, ...(f[element] || {}) };
+    }, {} as GraphQLSortArg);
+}
